fix(CB2-16150): handle different error responses
